+++ CRA_TeamProject_SSD/CRA_TeamProject_SSD.vcxproj.filters
@@ -134,5 +134,8 @@
     <ClInclude Include="ShellCommand\FlushCommand.h">
       <Filter>헤더 파일\ShellCommand</Filter>
     </ClInclude>
+    <ClInclude Include="Logger.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
